+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Master
 
-* Update `NavigationDropdown` to use accessible menu.
+- Update `NavigationDropdown` to use accessible menu.
   - 🚨`onDropdownChange` prop for `NavigationAccordion` and `onChange` prop for `NavigationDropdown` is now defunct and should be removed. This function is set automatically in `NavigationDropdown`.
   - 🚨`dropdownOptions` prop for `NavigationDropdown` is now `options`.